// src/shared/dataprovider/api/model/balanceTransferListDTOV2.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AgentBalanceTransfer } from './agentBalanceTransfer';

export interface BalanceTransferListDTOV2 { 
    Status?: AgentBalanceTransfer;
    StatusDescription?: string;
    TimeStamp?: Date;
    RequestID?: string;
    ClientId?: number;
    ClientName?: string;
    PaymentChannelName?: string;
    TotalAmount?: number;
    AgentId?: number;
    Requestee?: string;
    ApproverName?: string;
    DateCompleted?: Date;
    Remarks?: string;
    RequestedByAdmin?: string;
}