// src/shared/dataprovider/api/model/balanceTransferListDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BalanceTransferListDTOV2 } from './balanceTransferListDTOV2';
import { Pagination } from './pagination';

export interface BalanceTransferListDTO { 
    Paginations?: Pagination;
    Data?: Array<BalanceTransferListDTOV2>;
}